Major improvement of code (run-time algorithm and cleaning code). Add cryptanalysis folder and raifence tests + Gronsfeld + VigenereMult.

+++ src/Fleissner.hpp
@@ -12,23 +12,23 @@
 
 class Fleissner : public StringCipher
 {  
-   typedef std::pair<uint16_t, uint16_t> Cell;
+   typedef std::pair<uint32_t, uint32_t> Cell;
    typedef std::vector<Cell> Coordinates;
    using KeyCellNotUnique = BadKey;
    
 public:
-   Fleissner(const Coordinates &key, const uint16_t grid_dim, const bool clockwise);
+   Fleissner(const Coordinates &key, const uint32_t grid_dim, const bool clockwise);
    
    virtual const ClassicalType encode(const ClassicalType &clear_text) final;
    virtual const ClassicalType decode(const ClassicalType &cipher_text) final;
 
-   void setGridDimension(const uint16_t dim);
+   void setGridDimension(const uint32_t dim);
 
 private:   
-   void rotationExists(std::set<Cell> &cmp, const uint16_t x, const uint16_t y) const;
+   void rotationExists(std::set<Cell> &cmp, const uint32_t x, const uint32_t y) const;
 
    // Dimension of the cipher grid.
-   uint16_t grid_dim;
+   uint32_t grid_dim;
 
    Coordinates key;
 };
